requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
numpy==2.1.2
pillow==10.4.0
scikit-learn==1.5.2
rasterio==1.3.10
rich==13.9.2
matplotlib==3.9.2
pytest==8.3.3
